# Assorted performance optimizations for langtype and related code