--- a/internal/compiler/langtype.rs
+++ b/internal/compiler/langtype.rs
@@ -52,16 +52,7 @@ pub enum Type {
     Brush,
     /// This is usually a model
     Array(Box<Type>),
-    Struct {
-        fields: BTreeMap<SmolStr, Type>,
-        /// When declared in .slint as  `struct Foo := { }`, then the name is "Foo"
-        /// When there is no node, but there is a name, then it is a builtin type
-        name: Option<SmolStr>,
-        /// When declared in .slint, this is the node of the declaration.
-        node: Option<syntax_nodes::ObjectType>,
-        /// derived
-        rust_attributes: Option<Vec<SmolStr>>,
-    },
+    Struct(Rc<Struct>),
     Enumeration(Rc<Enumeration>),
 
     /// A type made up of the product of several "unit" types.
@@ -106,8 +97,8 @@ impl core::cmp::PartialEq for Type {
             Type::Easing => matches!(other, Type::Easing),
             Type::Brush => matches!(other, Type::Brush),
             Type::Array(a) => matches!(other, Type::Array(b) if a == b),
-            Type::Struct { fields, name, node: _, rust_attributes: _ } => {
-                matches!(other, Type::Struct{fields:f,name:n,node:_, rust_attributes: _ } if fields == f && name == n)
+            Type::Struct(lhs) => {
+                matches!(other, Type::Struct(rhs) if lhs.fields == rhs.fields && lhs.name == rhs.name)
             }
             Type::Enumeration(lhs) => matches!(other, Type::Enumeration(rhs) if lhs == rhs),
             Type::UnitProduct(a) => matches!(other, Type::UnitProduct(b) if a == b),
@@ -166,15 +157,17 @@ impl Display for Type {
             Type::Bool => write!(f, "bool"),
             Type::Model => write!(f, "model"),
             Type::Array(t) => write!(f, "[{}]", t),
-            Type::Struct { name: Some(name), .. } => write!(f, "{}", name),
-            Type::Struct { fields, name: None, .. } => {
-                write!(f, "{{ ")?;
-                for (k, v) in fields {
-                    write!(f, "{}: {},", k, v)?;
+            Type::Struct(t) => {
+                if let Some(name) = &t.name {
+                    write!(f, "{}", name)
+                } else {
+                    write!(f, "{{ ")?;
+                    for (k, v) in &t.fields {
+                        write!(f, "{}: {},", k, v)?;
+                    }
+                    write!(f, "}}")
                 }
-                write!(f, "}}")
             }
-
             Type::PathData => write!(f, "pathdata"),
             Type::Easing => write!(f, "easing"),
             Type::Brush => write!(f, "brush"),
@@ -281,9 +274,7 @@ impl Type {
             | (Type::Percent, Type::Float32)
             | (Type::Brush, Type::Color)
             | (Type::Color, Type::Brush) => true,
-            (Type::Struct { fields: a, .. }, Type::Struct { fields: b, .. }) => {
-                can_convert_struct(a, b)
-            }
+            (Type::Struct(a), Type::Struct(b)) => can_convert_struct(&a.fields, &b.fields),
             (Type::UnitProduct(u), o) => match o.as_unit_product() {
                 Some(o) => unit_product_length_conversion(u.as_slice(), o.as_slice()).is_some(),
                 None => false,
@@ -801,6 +792,18 @@ pub struct Callback {
 pub struct Function {
     pub return_type: Type,
     pub args: Vec<Type>,
+}
+
+#[derive(Debug, Clone)]
+pub struct Struct {
+    pub fields: BTreeMap<SmolStr, Type>,
+    /// When declared in .slint as  `struct Foo := { }`, then the name is "Foo"
+    /// When there is no node, but there is a name, then it is a builtin type
+    pub name: Option<SmolStr>,
+    /// When declared in .slint, this is the node of the declaration.
+    pub node: Option<syntax_nodes::ObjectType>,
+    /// derived
+    pub rust_attributes: Option<Vec<SmolStr>>,
 }
 
 #[derive(Debug, Clone)]